refactor(core): fold block metadata resolver into async_query
