+++ crates/minibf/src/routes/utxos.rs
@@ -4,7 +4,7 @@ use pallas::ledger::traverse::MultiEraOutput;
 use std::collections::{HashMap, HashSet};
 
 use dolos_cardano::indexes::AsyncCardanoQueryExt;
-use dolos_core::block_meta::BlockMetaResolver;
+use dolos_core::async_query::BlockMetaResolver;
 use dolos_core::{Domain, StateStore as _, TxHash, TxoIdx, TxoRef};
 
 use crate::{
