+++ crates/minikupo/src/routes/matches.rs
@@ -5,7 +5,7 @@ use axum::{
     Json,
 };
 use dolos_cardano::{indexes::CardanoStateIndexExt, network_from_genesis, pallas_extras};
-use dolos_core::block_meta::BlockMetaResolver;
+use dolos_core::async_query::BlockMetaResolver;
 use dolos_core::{ArchiveStore as _, Domain, EraCbor, StateStore as _, TxoRef, UtxoSet};
 use pallas::codec::minicbor;
 use pallas::ledger::{
